+++ lib/torch/nn/loss.rb
@@ -1,0 +1,9 @@
+module Torch
+  module NN
+    class Loss < Module
+      def initialize(reduction)
+        @reduction = reduction
+      end
+    end
+  end
+end