+++ lib/torch.rb
@@ -45,6 +45,7 @@ require "torch/nn/relu"
 require "torch/nn/sequential"
 
 # nn loss
+require "torch/nn/cross_entropy_loss"
 require "torch/nn/l1_loss"
 require "torch/nn/mse_loss"
 require "torch/nn/nll_loss"