+++ examples/mnist/main.rb
@@ -28,7 +28,7 @@ class Net < Torch::NN::Module
     x = Torch::NN::F.relu(x)
     x = @dropout2.call(x)
     x = @fc2.call(x)
-    output = Torch::NN::F.log_softmax(x, 1)
+    output = Torch::NN::F.log_softmax(x)
     output
   end
 end